+++ dev/app/builtin/stories/geo-search.stories.js
@@ -64,6 +64,30 @@ export default () => {
         start();
       })
     )
+  ).add(
+    'with transformed items',
+    wrapWithHitsAndConfiguration((container, start) =>
+      injectGoogleMaps(() => {
+        container.style.height = '600px';
+
+        window.search.addWidget(
+          instantsearch.widgets.geoSearch({
+            googleReference: window.google,
+            container,
+            transformItems: items =>
+              items.map(item => ({
+                ...item,
+                _geoloc: {
+                  lat: item._geoloc.lat + 2,
+                  lng: item._geoloc.lng + 2,
+                },
+              })),
+          })
+        );
+
+        start();
+      })
+    )
   );
 
   // With IP
@@ -201,17 +225,17 @@ export default () => {
     'with position from Places',
     wrapWithHitsAndConfiguration((container, start) =>
       injectGoogleMaps(() => {
-        const placesElemeent = document.createElement('input');
+        const placesElement = document.createElement('input');
         const mapElement = document.createElement('div');
         mapElement.style.height = '500px';
         mapElement.style.marginTop = '20px';
 
-        container.appendChild(placesElemeent);
+        container.appendChild(placesElement);
         container.appendChild(mapElement);
 
         window.search.addWidget(
           instantsearchPlacesWidget({
-            container: placesElemeent,
+            container: placesElement,
             defaultPosition: [position.lat, position.lng],
           })
         );