+++ dev/app/vanilla/widgets/index.js
@@ -1,6 +1,0 @@
-export { default as clearAll } from './clearAll';
-export { default as hits } from './hits';
-export { default as searchBox } from './searchBox';
-export { default as searchBoxReturn } from './searchBoxReturn';
-export { default as refinementList } from './refinementList';
-export { default as selectMenu } from './selectMenu';